[refactor] #12 - add tests
